tests/fixtures/python/example-auth0/requirements.txt
flask
python-dotenv
authlib
requests
